+++ Changelog.md
@@ -473,6 +473,7 @@
 - Reworked the part of the interface at the auxiliary rules pages for "other options and special flags".
 - The configuration directive for statistics has been split in two (tracking statistical information for CAPTCHAs now has its own configuration directive, separate from the other statistics), and the configuration for tracking statistical information for CAPTCHAs now distinguishes between platforms and can also track served CAPTCHAs (#624).
 - Dropped the "don't log" action from the auxiliary rules system. As the "suppress logging" option and the "don't log" action do the same thing, having both isn't necessary, so having users just use the "suppress logging" option is recommended instead (#334, #624).
+- Configuration directive `ban_override` removed. Configuration directive `http_response_header_code` reworked as a matrix, now able to set different HTTP response codes depending on whether a request was blocked due to a ban, due to legal reasons, or due to other reasons (#624).
 
 #### Bugs fixed.
 - Fixed various accessibility issues affecting the auxiliary rules pages.